% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{laraReport}
\alias{laraReport}
\alias{laraReport-package}
\title{A library for reporting data generated by lab automation systems}
\description{
A library for reporting data generated by lab automation systems
}
\details{
A library for reporting data generated by lab automation systems
}
\note{
date: 150601

last modification:
}
\author{
mark doerr \email{mark.doerr@uni-greifswald.de},
  and \email{mark@ismeralda.org}
}

